Fold over state on every input block
Previously, whatever the state was, it was only folded over when there
was an event. But this will not be the case for epoch stakepool size
indexer where the state is *ledger state*, which does change at every
block -- not just on blocks at which are at the epoch start.