Fix use of sqlite checkpoint for stateless indexers
Stateless indexers can start producing valid events at any chainpoint
so don't throw an exception when. They also keep a checkpoint in the
checkpoint table so one could resume. So, don't throw an exception when
the requested starting point is later than what was found in the
`checkpoints` table because the events will still be valid (the
indexer is stateless -- it can be started from anywhere). Skipping a
section is the responsibility of the user.