Add ability to index chain sections in parallel
This has been a major effort to attempt to index sections of a chain
in parallel in multiple threads. Indexers with no fold state can
benefit from it immediately, and some stateful indexers might benefit
in the future.

Though, preliminary tests have shown the speedup to be much less than I
had hoped it to be, will try to research why that is in the future.

By default all indexers are still run serially, parallelism can be
requested with the `--parallel` CLI switch.

There are now also two Cabal executable components, the previous
`mafoc` and a new `mafoc-threaded` which adds ghc options that create
a threaded executable.